Added the drone pickup logic
Created a scene for debris and code for the Carryable class, which allows them to be carried by drones and change aspects about them when they are, like scaling them up when a drone is hovering over it. I also created 2 new states for the drones, one for following the mouse and one for picking up debris. I created a new system for selecting a drone, and you can change which drone is selected by mousewheel scrolling up and down. The selected drone is the drone that will follow the mouse when you hold down the left mouse button. If you hover over a carryable object and then let go of LMB, your drone will enter a pickup state and pick up the object, holding them together and preventing them from picking up any other objects, as well as preventing that object from being picked up.