Did alot
I forgot about this after the first day :P
I set up the tilemap collisions and basics of the tileset terrains, the terrains are a bit broken but it is what it is. Set up procedurally generating map and randomly spawning map with cool resources and resource piles from before. Set up the shop and added a buncha assets